rename password validation component
