+++ app/(ui)/register/password/components/set-register-password-form.tsx
@@ -1,13 +1,13 @@
 "use client";
+import { useState } from "react";
+import { useTranslation } from "@i18n";
+import { validateAccount } from "@lib/validationSchemas";
 import { registerUser } from "@lib/server/register";
 import { PasswordComplexitySettings } from "@zitadel/proto/zitadel/settings/v2/password_settings_pb";
 import { useRouter } from "next/navigation";
 
-import { PasswordCreation } from "@clientComponents/forms/PasswordCreation/PasswordCreation";
+import { PasswordForm } from "@clientComponents/forms/PasswordValidation/PasswordForm";
 import { Alert, ErrorStatus } from "@clientComponents/forms";
-import { useState } from "react";
-import { useTranslation } from "@i18n";
-import { validateAccount } from "@lib/validationSchemas";
 
 export function SetRegisterPasswordForm({
   passwordComplexitySettings,
@@ -29,6 +29,8 @@ export function SetRegisterPasswordForm({
   const [error, setError] = useState("");
 
   const successCallback = async ({ password }: { password: string }) => {
+    // TODO validate password again
+
     // Validate account data again to be safe
     const validateAccountData = await validateAccount({ firstname, lastname, email } as {
       [k: string]: FormDataEntryValue;
@@ -59,7 +61,7 @@ export function SetRegisterPasswordForm({
   return (
     <>
       {error && <Alert type={ErrorStatus.ERROR}>{error}</Alert>}
-      <PasswordCreation
+      <PasswordForm
         passwordComplexitySettings={passwordComplexitySettings}
         successCallback={successCallback}
       />
